Removing the queue from ingestCtrl (simplifying overengineering xd) - Adding config sensor resending logic

+++ apps/api/src/lib/controllers/ingestCtrl.ts
@@ -1,34 +1,19 @@
 import { RequestHandler } from "express";
 import { IMessageBus } from "../sockets/messageBus";
-import { IQueue } from "../queue/IQueue";
 import { ConfigData, ReadingData } from "../../types/sensorTypes";
 import SensorRepository from "../repositories/sensorRepository";
 import RedisRepository from "../repositories/cache/redisRepository";
 import PostgresRepository from "../repositories/database/postgresRepository";
-import Queue from "../queue/queueConcration";
 import { Reading } from "../../database/mongo.db";
 import { createFormattedInfoInfo } from "../services/infoFormatterService";
 
 const cacheRepository = new RedisRepository();
-const sensorRepository = new SensorRepository(cacheRepository, new PostgresRepository());
+const sensorRepository = new SensorRepository(
+    cacheRepository,
+    new PostgresRepository()
+);
 
 export function createIngestCtrl(bus: IMessageBus): RequestHandler {
-    const queue: IQueue<ReadingData> = new Queue<ReadingData>(100);
-
-    queue.setHandler(async (data: ReadingData) => {
-        const reading = new Reading(data);
-        await reading.save();
-
-        const readingSensorConfig: ConfigData | null = await sensorRepository.getSensorConfig(data.sensorId);
-        if (!readingSensorConfig) {
-            console.warn(`⚠️ No sensor config found for sensorId: ${data.sensorId}`);
-            return;
-        }
-
-        const info = await createFormattedInfoInfo(data, readingSensorConfig);
-        bus.publish("reading", info);
-    });
-
     return async (req, res) => {
         try {
             const data: ReadingData = req.body;
@@ -39,18 +24,29 @@ export function createIngestCtrl(bus: IMessageBus): RequestHandler {
             if (!data.metrics || !data.sensorId || !data.timestamp)
                 return res.status(400).json({ error: "Missing required fields" });
 
+            const reading = new Reading(data);
+            await reading.save();
+
+            const readingSensorConfig: ConfigData | null =
+                await sensorRepository.getSensorConfig(data.sensorId);
+
+            if (!readingSensorConfig) {
+                console.warn(`⚠️ No sensor config found for sensorId: ${data.sensorId}`);
+                return res.status(404).json({ ok: false });
+            }
+
+            const info = await createFormattedInfoInfo(data, readingSensorConfig);
+            bus.publish("reading", info);
+
             const newConfig = await cacheRepository.get(`sensor:${data.sensorId}-updated`);
             if (newConfig) {
                 const parsed = JSON.parse(newConfig || "null");
 
-                // Limpiamos las keys relacionadas
                 await cacheRepository.delete(`sensor:${data.sensorId}-updated`);
                 await cacheRepository.delete(sensorRepository.getCacheKey(data.sensorId));
 
                 return res.status(202).json({ ok: true, config: parsed });
             }
-
-            queue.add(data);
 
             return res.status(202).json({ ok: true });
         } catch (err) {
@@ -66,6 +62,7 @@ export const addSensorCtrl: RequestHandler = async (req, res) => {
 
     try {
         let sensor = await sensorRepository.getSensorConfig(data.sensorId);
+
         if (sensor) {
             console.log(`✅ Config encontrada para ${sensor.sensorId}`);
             return res.status(201).json({ ok: true, sensor });
